Updating Head Matomo TM code from GTM Part 2
Took out the 'if statements' from line 2 and line 12 'the if jekyll.environment == "production' as requested. Replaced the Matomo code with the dev Matomo code from Library Discovery Matomo Tag Manager page- clicking on Install code on that page and selecting dev from the drop down menu at the top of the page.  Copy and pasted that code here.